refactor: src/features/canvas -> src/features/editor 폴더이름 변경
