+++ src/app/router/routes.tsx
@@ -1,6 +1,6 @@
 import { createBrowserRouter } from "react-router-dom";
 import HomePage from "../../pages/home_page";
-import { MathChatInput } from "../../features/canvas/components/MathChatInput";
+import { MathChatInput } from "../../features/editor/components/MathChatInput";
 
 export const router = createBrowserRouter([
   {
@@ -8,7 +8,7 @@ export const router = createBrowserRouter([
     element: <HomePage />,
   },
   {
-    path: "/canvas",
+    path: "/editor",
     element: (
       <div className="flex h-screen w-full items-center justify-center bg-gray-50 p-10">
         <MathChatInput />
