+++ src/features/editor/components/MathChatInput.tsx
