+++ src/features/editor/components/input/ChatInputArea.tsx
