+++ src/features/editor/components/layer/MathLayer.tsx
